# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Veris Helpdesk API
  version: 1.0.0
  description: Emails are compared case-insensitively everywhere, including OTP and ticket matching.
servers:
  - url: /
paths:
  /api/otp/request:
    post:
      summary: Request a one-time verification code
      operationId: requestOtp
      security: []
      description: >
        Always returns the same success response regardless of whether the
        email has tickets. A code is sent only when the email has at least one
        ticket. Emails are compared case-insensitively.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: Identical success response for all email addresses.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
                    enum: ["If this email has tickets, a code was sent."]
                    example: "If this email has tickets, a code was sent."
        "400":
          $ref: "#/components/responses/BadInput"
        "429":
          description: >
            Rate limit exceeded. A resend requested within 60 seconds returns
            OTP_RESEND_WAIT and retryAfterSeconds; a generic rate limit has no code.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericRateLimitError"
                  - $ref: "#/components/schemas/OtpResendWaitError"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session
      description: >
        Administrator JWT cookie. Protected admin endpoints require a token
        with an adminId payload. Client tokens with role: client must be rejected.
    clientSession:
      type: apiKey
      in: cookie
      name: client_session
      description: >
        Client JWT cookie; payload is { email, role: "client" }. Admin endpoints
        require an admin token with adminId; client tokens must be rejected.
...
  responses:
    BadInput:
      description: Bad input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServerError:
      description: Server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
    GenericRateLimitError:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
    OtpResendWaitError:
      type: object
      required: [error, code, retryAfterSeconds]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [OTP_RESEND_WAIT]
        retryAfterSeconds:
          type: integer
...
    OtpRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...